Project update
Hi, this project started with the idea of having multiple languages and libraries. However I've decided to change this idea, from now on I'll have a project focused on a single library or language.

The project is focused on cypress now, I've expanded the login test including api request validations and testing up to a full login. No real information has been used in the tests.