Expanded on the SignUp test.
I've add more lines for testing the signup page, up to the name and lastname fields. It was troublesome to deal with the selectors, but I've found two usable ones:
jsname;
id;

Implemented the cypress library, since it was missing in the last commit, now the code is testable and working.